Fix 10 low-risk bugs: RadarAlt deinit order, IFEI overflow guard, wifi timeout, strcmp, dead code cleanup
- I6: Swap vTaskDelete/waitDMADone order in RadarAlt_deinit to prevent SPI corruption
- I7: Add sizeof bounds check to IFEIPanel copyN for tempL/tempR buffers
- I20: Add 20s timeout to wifi_setup() connection loop to prevent infinite hang
- M3: Use strcmp instead of pointer comparison in queueDeferredRelease
- M6: Remove dead hc165Bits static variable from Generic.cpp (only hc165PrevBits used in loop)
- M9: Widen RS485 skipRemaining from uint8_t to uint16_t to prevent overflow on c >= 254
- M11: Add bounds check to displayedIndexes array in LED test menu
- M16: Remove dead HIDSenderTask and shouldPollDisplayRefreshMs declarations from HIDManager.h
- M17: Change #if CLOSE_CDC_SERIAL to #if defined(CLOSE_CDC_SERIAL) for consistency
- M18: Add missing newline to wifiDebugPrintln to match println semantics

+++ lib/CUtils/src/internal/RS485Slave.cpp
@@ -174,7 +174,7 @@ static volatile uint8_t packetLength = 0;
 static volatile uint16_t packetDataIdx = 0;
 
 // Skip state
-static volatile uint8_t skipRemaining = 0;
+static volatile uint16_t skipRemaining = 0;
 
 // Data type for filtering what goes into export buffer
 enum RxDataType : uint8_t {
